moved ProjektionMidiClass TestHelper(Macros) to own file
